% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assocsim-package.R
\docType{package}
\name{assocsim-package}
\alias{assocsim}
\alias{assocsim-package}
\title{assocsim: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
