% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2sparseMatrix.R
\name{conversions-list-sparseMatrix}
\alias{conversions-list-sparseMatrix}
\alias{list2sparseMatrix}
\alias{sparseMatrix2list}
\title{Convert between list and sparse matrix representations}
\usage{
list2sparseMatrix(x)

sparseMatrix2list(m)
}
\arguments{
\item{x}{A list}

\item{m}{A sparseMatrix}
}
\value{
The same data in the complimentary format
}
\description{
Convert between list and sparse matrix representations
}
\details{
Lists are converted into "C" sparse matrices.
}
