% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine-similarity.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Compute cosine similarity matrix}
\usage{
cosine_similarity(x, ...)
}
\arguments{
\item{x}{A list of named vectors, a matrix, or a sparseMatrix}

\item{y}{A vector; `\code{y}` will be ignored or rejected if \code{x} is not
vector.}
}
\description{
Similarities are computed among rows or list elements. When \code{x} is a
list, it is assumed that each element contains a named vector.
}
