% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\alias{normalize_rows}
\alias{normalize_cols}
\alias{normalize_rows.list}
\title{Normalize data structures by rows or columns}
\usage{
normalize_rows(x, norm = c("l1", "l2"))

normalize_cols(x, norm = c("l1", "l2"))

\method{normalize_rows}{list}(x, norm = c("l1", "l2"))
}
\arguments{
\item{x}{A list of numeric vectors, a matrix, or a sparseMatrix}

\item{norm}{The norm to apply to each row or column. Either "l1" to make sum
of row = 1, or "l2" to make sum of squared row = 1.}
}
\value{
Normalized data, structured as the input \code{x} was structured.
}
\description{
Normalize data structures by rows or columns
}
