% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positive-pointwise-mutal-information.R
\name{ppmi}
\alias{ppmi}
\title{Positive pointwise mutual information for responses conditioned on cues}
\usage{
ppmi(x, ...)
}
\arguments{
\item{x}{A data.frame, list, matrix, or sparseMatrix containing association
strengths.}

\item{value_var, cue_var, resp_var}{If \code{x} is a data.frame, the names
of these three critical columns need to be provided.}

\item{rprob}{The mean response strength over cues (i.e., average probability
of each response, where each response has been conditioned on all cues).}
}
\description{
Positive pointwise mutual information for responses conditioned on cues
}
\details{
If supplying \code{rprob}, it should be computed as follows:
1. Let X be a matrix, with rows corresponding to cues and columns to responses.
2. Normalize each row of X to sum to 1.
3. Compute column means of X. This vector is \code{rprob}.

You may wish to compute \code{rprob} based on a larger, normative dataset to
more accurately account for the overall probability of each response over a
large set of cues.
}
