% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{similarity}
\alias{similarity}
\title{Class for representing symmetric similarity matrix as lower-triangle vector}
\usage{
similarity(x)
}
\arguments{
\item{x}{A symmetric similarity matrix}
}
\description{
This is a simple adaptation of the \code{dist} S3 class
}
