% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_of_cues.R
\name{strong_component}
\alias{strong_component}
\title{Identify and extract the largest strong component from a network}
\usage{
strong_component(g)
}
\arguments{
\item{g}{An igraph representation of a simple network of cues}
}
\value{
The largest set of nodes with in- and out-degree greater than 0.
}
\description{
Identify and extract the largest strong component from a network
}
